# SYCL: Introducing memory host pool