--- a/ggml/src/ggml-sycl/common.hpp
+++ b/ggml/src/ggml-sycl/common.hpp
@@ -352,14 +352,12 @@ struct ggml_backend_sycl_context {
 
     ggml_sycl_pool & host_pool(int device) {
         if (host_pools[device] == nullptr) {
-            host_pools[device] = new_pool_for_host(stream(device,0), device);
+            host_pools[device] = new_pool_for_host(stream(device, 0), device);
         }
         return *host_pools[device];
     }
 
-    ggml_sycl_pool & host_pool() {
-        return host_pool(device);
-    }
+    ggml_sycl_pool & host_pool() { return host_pool(device); }
 };
 
 // common device functions