- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
 - Self-reported review complexity:
     - [x] Low
     - [ ] Medium
     - [ ] High

This patch adds a host memory pool dedicated to `matrix_info_t`. This struct is used when calling `gemm_batch_impl` and right now forces an `host_task` synchronization to free memory. After this patch it is not necessary anymore. The pool ensures that the memory will be usable long enough to call the kernel and it will be freed at the end of the program. 

Pool's design is naive and right now usable only with above reported type, further improvement are possible but not necessary at the moment.

Full test pass on A100 and Intel PVC.
Benchmark results for NVIDIA:

### FP32
**Current**

| model                          |       size |     params | backend    | ngl | threads |    sm | mmap |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------: | ----: | ---: | ------------: | -------------------: |
| llama 8B Q8_0                  |   7.95 GiB |     8.03 B | SYCL       |  99 |       8 |  none |    0 |         pp512 |        759.05 ± 4.70 |
| llama 8B Q8_0                  |   7.95 GiB |     8.03 B | SYCL       |  99 |       8 |  none |    0 |         tg128 |         89.10 ± 0.16 |
| llama 70B Q4_K - Small         |  37.57 GiB |    70.55 B | SYCL       |  99 |       8 |  none |    0 |         pp512 |         72.93 ± 0.31 |
| llama 70B Q4_K - Small         |  37.57 GiB |    70.55 B | SYCL       |  99 |       8 |  none |    0 |         tg128 |         18.17 ± 0.02 |
| llama 8B Q4_K - Medium         |   4.58 GiB |     8.03 B | SYCL       |  99 |       8 |  none |    0 |         pp512 |        743.26 ± 1.85 |
| llama 8B Q4_K - Medium         |   4.58 GiB |     8.03 B | SYCL       |  99 |       8 |  none |    0 |         tg128 |         91.47 ± 0.07 |

build: a29f0870 (4473)

**This patch**


| model                          |       size |     params | backend    | ngl | threads |    sm | mmap |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------: | ----: | ---: | ------------: | -------------------: |
| llama 8B Q8_0                  |   7.95 GiB |     8.03 B | SYCL       |  99 |       8 |  none |    0 |         pp512 |        969.44 ± 7.14 |
| llama 8B Q8_0                  |   7.95 GiB |     8.03 B | SYCL       |  99 |       8 |  none |    0 |         tg128 |         89.98 ± 0.18 |
| llama 70B Q4_K - Small         |  37.57 GiB |    70.55 B | SYCL       |  99 |       8 |  none |    0 |         pp512 |        104.87 ± 0.21 |
| llama 70B Q4_K - Small         |  37.57 GiB |    70.55 B | SYCL       |  99 |       8 |  none |    0 |         tg128 |         18.27 ± 0.01 |
| llama 8B Q4_K - Medium         |   4.58 GiB |     8.03 B | SYCL       |  99 |       8 |  none |    0 |         pp512 |        957.31 ± 2.29 |
| llama 8B Q4_K - Medium         |   4.58 GiB |     8.03 B | SYCL       |  99 |       8 |  none |    0 |         tg128 |         92.51 ± 0.05 |

build: e878c29a (4476)

### FP16

**Current**


| model                          |       size |     params | backend    | ngl | threads |    sm | mmap |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------: | ----: | ---: | ------------: | -------------------: |
| llama 8B Q8_0                  |   7.95 GiB |     8.03 B | SYCL       |  99 |       8 |  none |    0 |         pp512 |      5557.98 ± 18.23 |
| llama 8B Q8_0                  |   7.95 GiB |     8.03 B | SYCL       |  99 |       8 |  none |    0 |         tg128 |         88.98 ± 0.19 |
| llama 70B Q4_K - Small         |  37.57 GiB |    70.55 B | SYCL       |  99 |       8 |  none |    0 |         pp512 |        694.68 ± 1.90 |
| llama 70B Q4_K - Small         |  37.57 GiB |    70.55 B | SYCL       |  99 |       8 |  none |    0 |         tg128 |         18.15 ± 0.02 |
| llama 8B Q4_K - Medium         |   4.58 GiB |     8.03 B | SYCL       |  99 |       8 |  none |    0 |         pp512 |      5375.49 ± 23.61 |
| llama 8B Q4_K - Medium         |   4.58 GiB |     8.03 B | SYCL       |  99 |       8 |  none |    0 |         tg128 |         91.56 ± 0.06 |

build: a29f0870 (4473)

**This patch**


| model                  |      size |  params | backend | ngl | threads |   sm | mmap |  test |             t/s |
| ---------------------- | --------: | ------: | ------- | --: | ------: | ---: | ---: | ----: | --------------: |
| llama 8B Q8_0          |  7.95 GiB |  8.03 B | SYCL    |  99 |       8 | none |    0 | pp512 | 5606.51 ± 79.11 |
| llama 8B Q8_0          |  7.95 GiB |  8.03 B | SYCL    |  99 |       8 | none |    0 | tg128 |    87.32 ± 6.34 |
| llama 70B Q4_K - Small | 37.57 GiB | 70.55 B | SYCL    |  99 |       8 | none |    0 | pp512 |   707.81 ± 3.43 |
| llama 70B Q4_K - Small | 37.57 GiB | 70.55 B | SYCL    |  99 |       8 | none |    0 | tg128 |    18.35 ± 0.06 |
| llama 8B Q4_K - Medium |  4.58 GiB |  8.03 B | SYCL    |  99 |       8 | none |    0 | pp512 | 5404.64 ± 16.42 |
| llama 8B Q4_K - Medium |  4.58 GiB |  8.03 B | SYCL    |  99 |       8 | none |    0 | tg128 |    92.62 ± 0.07 |

build: e878c29a (4476)