@s-Nick 2025-01-15
Thank you for the comment @qnixsynapse  I address it in [6b77639](https://github.com/ggerganov/llama.cpp/pull/11251/commits/6b776392586bfc6c16cee171c2b153812a9062f6)