@qnixsynapse 2025-01-15
@s-Nick Sorry for being a bit fussy about compiler warnings. There is one tiny issue left which can be fixed by adding a `return nullptr;` before the closing brackets of the alloc function which I have put out a comment.

So far everything else LGTM.