@s-Nick 2025-01-16
@qnixsynapse  don't worry, more eyes are always better, thank you for reviewing it. I addressed it in [963b685](https://github.com/ggerganov/llama.cpp/pull/11251/commits/963b6850751ef6a17c4c58eb6b20d2226fd1ea6c)