@qnixsynapse 2025-01-16
@s-Nick Thanks. LGTM!